Make Options type public
I'm writing a wrapper for hotkeys in TypeScript, and it receives options as an argument. I'd like to type that argument but I can't because it's not public. Unless there's a particular reason for it to be private, it should be public since there are valid use cases for it.